put components into components folder
