+++ client/src/App.js
@@ -5,9 +5,9 @@ import logo from './logo.png';
 // import Scoreboard from './components/Scoreboard';
 import { Nav, Navbar, Button, Container, Row, Col } from 'react-bootstrap';
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
-import Home from "./home"
-import Header from "./header"
-import join_room from "./join_room";
+import Home from "./components/home"
+import Header from "./components/header"
+import join_room from "./components/join_room";
 
 const App = () => (
 <Router>
