+++ client/src/components/home.js
@@ -1,33 +1,24 @@
 import React from 'react';
-import './App.css';
-import logo from './logo.png';
+import '../App.css';
+import logo from '../logo.png';
 // import GenButton from './components/Button';
 // import Scoreboard from './components/Scoreboard';
 import { Nav, Navbar, Button, Container, Row, Col } from 'react-bootstrap';
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 import join_room from "./join_room";
+import Header from "./header";
 
-const Header = () => (
+const Home = () => (
   <>
 
 
 
-
-      <Navbar fixed="top" bg="info" expand="lg">
-        <Navbar.Brand href="/">
-          <img alt="quip logo" src={logo} width="85" height="44.25" className="d-inline-block align-top"/>
-        </Navbar.Brand>
-        {/* <Nav.Link href="#" style={{color:'white'}}>About</Nav.Link> */}
-        <Container className="justify-content-end">
-          <Row>
-            <Nav.Link href="#"><Col><Button variant="primary">Create a Room</Button></Col></Nav.Link>
-            <Nav.Link href="/join_room"><Col><Button variant="success">Join a Room</Button></Col></Nav.Link>
-            <Nav.Link href="#"><Col><Button variant="danger">Add a Prompt</Button></Col></Nav.Link>
-          </Row>
-        </Container>
-      </Navbar>
-
-
+    <div className="App">
+      <Header/>
+      <div className="contentContainer heading" id="first">
+        <h1 id="title">Quip Thinking</h1>
+        <h4 id="byline">the online game of witty lines</h4>
+      </div>
 
       {/* <div className="contentContainer" id="second">
         <h2>Play now!</h2>
@@ -44,9 +35,9 @@ const Header = () => (
         <hr />
         <Scoreboard />
       </div> */}
-
+    </div>
 
   </>
 )
 
-export default Header;
+export default Home;
